Add new command Relative
This command permits to get the uuid of a linked resource, by
providing a resource and a resource type path. The resource type path
is a list of resource types separated by dots.

For instance, to get the logical router associated to a virtual
machine interface that implements a SNAT:
contrail-api-cli relative virtual-machine-interface/737cad4e-b4ca-4e67-a3f0-fb1996a63a05 virtual_machine.service_instance.logical_router

Be careful, when several children are available, the first one is always choosen.

Note this command supports 'child' attributes (such as
loadbalancer-pool.loadbalancer-members) but this is not yet
implemented in the core.